allow sending tracing metadata (target, file/line, module_path) as syslog structured data
this adds optional features to pass on more of the metadata from the tracing infrastructure
as structured-data, when using RFC-5424 format.

it doesn't go as far as sending fields and valuesets from tracing events, but still seems
like a significant improvement.

note that `tracing_subscriber` typically defaults to formatting the target
with the log message in text logs, so many users are used to seeing that in text logs,
but it's not present in the syslog records produced by this crate before this change.
in tracing_subscriber, the file and line number are optional defaulting to off.
in this PR we make all this stuff default to off, for backwards compatible behavior.