add optional integration with tracing-log
tracing-log is a bridge that allows tracing to pick up log messages
written by older rust libraries that used the `log` crate facade.
the bridge subscribes to log messages and resubmits them as tracing
events.

because of lifetime issues, it can't take the log message file and
line numbers and register them directly as file and line numbers of
the tracing events. instead it places them as `log.file` and `log.line`
in the event metadata. then, when using `tracing-subscriber`, the
metadata are "normalized" by repopulating the file and line number
information. This normalization is possible when `tracing-log` feature
is turned on.

in this commit we mirror this behavior, so that tracing-rfc-5424 will
also make log messages carry this metadata just as well as tracing messages,
if the optional `tracing-log` feature is turned on